Sunday Fixes
Fixed the WebConfig and Plugged in the User Registration Page as well as
the Log In page to the database.
