+++ MathFun1000/MathFun1000.csproj
@@ -213,7 +213,6 @@
     <Content Include="favicon.ico" />
     <Content Include="Global.asax" />
     <Content Include="LoggedIn.aspx" />
-    <Content Include="LogIn.aspx" />
     <Content Include="MathProgram.aspx" />
     <Content Include="packages.config" />
     <None Include="Properties\PublishProfiles\Dan.pubxml" />
@@ -245,7 +244,6 @@
     <Content Include="Scripts\WebForms\WebForms.js" />
     <Content Include="Scripts\WebForms\WebParts.js" />
     <Content Include="Scripts\WebForms\WebUIValidation.js" />
-    <Content Include="UserRegistration.aspx" />
     <Content Include="ViewSwitcher.ascx" />
     <Content Include="Web.config" />
     <Content Include="Bundle.config" />
@@ -347,7 +345,7 @@
       <DependentUpon>Graph.aspx</DependentUpon>
     </Compile>
     <Compile Include="Graphs.cs" />
-    <Compile Include="InputValidation.cs" />
+    <Compile Include="Account\InputValidation.cs" />
     <Compile Include="IProblemType.cs" />
     <Compile Include="LoggedIn.aspx.cs">
       <DependentUpon>LoggedIn.aspx</DependentUpon>
@@ -356,13 +354,6 @@
     <Compile Include="LoggedIn.aspx.designer.cs">
       <DependentUpon>LoggedIn.aspx</DependentUpon>
     </Compile>
-    <Compile Include="LogIn.aspx.cs">
-      <DependentUpon>LogIn.aspx</DependentUpon>
-      <SubType>ASPXCodeBehind</SubType>
-    </Compile>
-    <Compile Include="LogIn.aspx.designer.cs">
-      <DependentUpon>LogIn.aspx</DependentUpon>
-    </Compile>
     <Compile Include="MathProgram.aspx.cs">
       <DependentUpon>MathProgram.aspx</DependentUpon>
       <SubType>ASPXCodeBehind</SubType>
@@ -370,7 +361,7 @@
     <Compile Include="MathProgram.aspx.designer.cs">
       <DependentUpon>MathProgram.aspx</DependentUpon>
     </Compile>
-    <Compile Include="PasswordHash.cs" />
+    <Compile Include="Account\PasswordHash.cs" />
     <Compile Include="Problem.cs" />
     <Compile Include="Problems.aspx.cs">
       <DependentUpon>Problems.aspx</DependentUpon>
@@ -397,13 +388,6 @@
     <Compile Include="tester_class.cs" />
     <Compile Include="Tutorial.cs" />
     <Compile Include="Unguided.cs" />
-    <Compile Include="UserRegistration.aspx.cs">
-      <DependentUpon>UserRegistration.aspx</DependentUpon>
-      <SubType>ASPXCodeBehind</SubType>
-    </Compile>
-    <Compile Include="UserRegistration.aspx.designer.cs">
-      <DependentUpon>UserRegistration.aspx</DependentUpon>
-    </Compile>
     <Compile Include="ViewSwitcher.ascx.cs">
       <DependentUpon>ViewSwitcher.ascx</DependentUpon>
       <SubType>ASPXCodeBehind</SubType>
