+++ MathFun1000/Web.config
@@ -19,12 +19,13 @@
         path="ChartImg.axd" type="System.Web.UI.DataVisualization.Charting.ChartHttpHandler, System.Web.DataVisualization, Version=4.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35" />
     </handlers>
   </system.webServer>
-  <connectionStrings>
-    <add name ="webAppconnString"
+  <!--<connectionStrings>
+    <add name ="DefaultConnection"
          connectionString="Server=MYSQL5002.Smarterasp.net;Uid=9bad3d_test;Password=math4fun#;Database=db_9bad3d_test;"
          providerName="Math Fun 1000"/>
-  </connectionStrings>
+  </connectionStrings>-->
   <connectionStrings>
+    <add name="WebAppConnString" providerName="Math Fun 1000" connectionString="Server=MYSQL5002.Smarterasp.net;Uid=9bad3d_test;Password=math4fun#;Database=db_9bad3d_test;"/>
     <add name="DefaultConnection" providerName="System.Data.SqlClient" connectionString="Data Source=(LocalDb)\v11.0;Initial Catalog=aspnet-MathFun1000-20150127154319;Integrated Security=SSPI;AttachDBFilename=|DataDirectory|\aspnet-MathFun1000-20150127154319.mdf"/>
   </connectionStrings>
   <system.web>
